Big Update
Added lives and Respawning as well as update the function of walls.
Instead of instantly dying now it just keeps you at bay and takes away a
snakePart if you keep running into it. I like it a lot better. I really
want to get a menu screen and add options like lives, map, and players
on a GUI Screen.