Spring cleaning
Cleaned up most of the code a lot. Got rid of repetetive statements and
stuff like that. Also added a better color function to the snakes. When
a snake eats the other, it gains snakeParts that are that color. It's
pretty cool. Overall the code, I think, is much better and it would be
really easy to implement different screens and even AI. YAY!